+++ apps/api/src/api/products/(products)/kvstore/kvstore.controller.ts
@@ -1,7 +1,6 @@
 import { Controller, Post, Body, Delete, Query, BadRequestException, NotFoundException, Get } from "@nestjs/common"
 import { KvstoreService } from "./kvstore.service"
 import { CreateKvDto } from "./dto/create-kv.dto"
-import { TokenAuthorizer, TokenAuthorizerResponse } from "src/authorization/token-authorizer.decorator"
 import { CredentialAuthorizer, CredentialAuthorizerResponse } from "src/authorization/credential-authorizer.decorator"
 
 @Controller("products/kvstore")
@@ -19,20 +18,8 @@ export class KvstoreController {
     }
   }
 
-  @Get("readkvlistfromplatform")
-  async readKvListInsidePlatform(@TokenAuthorizer() uft: TokenAuthorizerResponse) {
-    try {
-      const { kvs } = await this.kvstoreService.readKvList(uft.workspaceId)
-      return { kvs }
-    }
-
-    catch (error) {
-      throw new NotFoundException()
-    }
-  }
-
-  @Get("readkvlist")
-  async readKvListOutsidePlatform(@CredentialAuthorizer() ufc: CredentialAuthorizerResponse) {
+  @Get("readkv")
+  async readKvList(@CredentialAuthorizer() ufc: CredentialAuthorizerResponse) {
     try {
       const { kvs } = await this.kvstoreService.readKvList(ufc.workspaceId)
       return { kvs }