+++ apps/api/src/authorization/credential-authorizer.decorator.ts
@@ -1,7 +1,7 @@
 import { createParamDecorator, ExecutionContext, ForbiddenException } from "@nestjs/common"
 import { findSubscriptionByUserIdQuery } from "src/api/subscription/queries/find-subscription"
 import { findWorkspaceByCredentialQuery } from "src/api/workspace/queries/find-workspace-by-credential.query"
-import { apiPricing, SubscriptionPlans } from "src/api/subscription/subscription.config"
+import { SubscriptionPlans } from "src/api/subscription/subscription.config"
 import { statusMessages } from "src/constants/status-messages"
 import { delay } from "src/lib/delay"
 
@@ -13,9 +13,8 @@ export interface CredentialAuthorizerResponse {
 export const CredentialAuthorizer = createParamDecorator(
   async (data: unknown, ctx: ExecutionContext): Promise<CredentialAuthorizerResponse> => {
     const request = ctx.switchToHttp().getRequest()
-    const clientId = request.headers["client_id"]
-    const clientSecret = request.headers["client_secret"]
-    const requestedResource = String(request.originalUrl).split("/")[3]
+    const clientId = request.headers["client_id"] || request.query.client_id
+    const clientSecret = request.headers["client_secret"] || request.query.client_secret
 
     if (!clientId || !clientSecret) {
       throw new ForbiddenException(statusMessages.noCredentialsProvided)
@@ -39,9 +38,9 @@ export const CredentialAuthorizer = createParamDecorator(
             }
 
             else {
-              const creditRequiredForCurrentRequest = apiPricing[`${requestedResource}`]
+              const creditRequired = 1
 
-              if (creditRequiredForCurrentRequest > subscription.remainingCredits) {
+              if (creditRequired > subscription.remainingCredits) {
                 throw new ForbiddenException(statusMessages.subscriptionLimitReached)
               }
 
@@ -58,7 +57,7 @@ export const CredentialAuthorizer = createParamDecorator(
                   await delay(100)
                 }
 
-                subscription.remainingCredits -= creditRequiredForCurrentRequest
+                subscription.remainingCredits -= creditRequired
                 await subscription.save()
                 return { userId, workspaceId }
               }