+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+### v11.2.0
+* Added Credential Authorizer in all tx gateways to enable API solution for blockchain apps 
+* Removed all Create TX API
+* Removed API Pricing Credits
+* Restructured Pricing & Credits 
+* Renamed Credits to API calls
+* New Analytics Get API released outside platform
+
 ### v11.1.0
 * Bug fixes & performance improvements
 * Unnecessary code cleanup